requirements.txt
Flask
Werkzeug
opencv-python-headless
numpy
tensorflow
imageio
Pillow
streamlit